move create project dto to DataTransfer project
