--- TicketPusher.Server/Projects/CreateProjectDto.cs
@@ -1,10 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace TicketPusher.Server.Projects
-{
-    public class CreateProjectDto
-    {
-        [Required]
-        public string Name { get; set; }
-    }
-}
