+++ TicketPusher.Server/Projects/IProjectWriteDataService.cs
@@ -1,3 +1,4 @@
+using TicketPusher.DataTransfer.Projects;
 using TicketPusher.Server.Templates;
 
 namespace TicketPusher.Server.Projects
