+++ TicketPusher.Server/TicketPusher.Server.csproj
@@ -9,6 +9,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\TicketPusher.Domain\TicketPusher.Domain.csproj" />
+    <ProjectReference Include="..\TicketPusher.DataTransfer\TicketPusher.DataTransfer.csproj" />
   </ItemGroup>
 
   <PropertyGroup>
